+++ NEWS.md
@@ -16,6 +16,7 @@ New features(Analysis):
 
 Maintenance
 + Don't emit a warning to stderr when `--language-server-completion-vscode` is used.
++ Catch the rare RecursionDepthException in more places, improve readability of its exception message. (#2386)
 
 Bug fixes:
 + Fix edge cases in checking if properties/methods are accessible from a trait (#2371)